PlayerStatusCardsコンポーネントを追加し、Mainコンポーネントでのプレイヤーステータス表示を改善
